did types structuring
