--- src/types/apiResoponse.ts
@@ -1,5 +1,0 @@
-export interface ApiResponse {
-    success: boolean;
-    message: string;
-}
-
