--- src/types/articlesResponse.ts
@@ -1,6 +1,0 @@
-import { Article } from '@/types/Article';
-
-export interface ArticlesResponse {
-  success: boolean;
-  articles: Article[];
-}
