--- src/types/navlinks.ts
@@ -1,9 +1,0 @@
-
-export interface NavLink {
-    name: string;
-    href: string;
-}
-
-export default interface NavLinks {
-    [key: string]: NavLink[];
-}
