+++ src/types/postObject.ts
@@ -1,15 +1,28 @@
-export default interface PostObject {
+import { Document } from "mongoose";
+
+// TODO: to remove this need to check
+interface Element {
+  type: string;
+  content?: string;
+  attrs?: Record<string, any>;
+}
+
+export default interface PostObject extends Document{
     id: string;
     title: string;
-    content: string;
-    type: string;
-    publisher: string;
-    date: string;
-    modified: string;
-    tags: string[];
-    event: string;
-    featured_image: string;
+    oneLiner?: string;
+    content?: string;
+    publishedIn: Date;
+    featuredImage?: string; // Path to the featured image
+    publisher: {
+        _id: string;
+        name: string;
+        username: string;
+    };
+    voteCount: number;
+    postTags: string[];
+    modifiedIn: Date;
+    trashed?: boolean;
     category: string;
-    author?: string; // Optional for news type
-    publisherID?: string;
+    link: string;
 }
