+++ src/types/post_objects/article.ts
@@ -1,0 +1,4 @@
+import PostObject from "../postObject";
+export default interface Article extends PostObject {
+  author: string;
+}
