+++ src/types/post_objects/notice.ts
@@ -1,0 +1,3 @@
+import PostObject from "../postObject";
+
+export default interface Notice extends PostObject {}
